pyproject.toml
[project]
name = "django-df-cards"
version = "0.0.1"
description = "django-df-cards"
readme = "README.md"
authors = [{name = "Apexive OSS", email = "open-source@apexive.com"}]
license = { file = "LICENSE" }
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 3.2",
    "Framework :: Django :: 4.0",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Topic :: Internet"
]
requires-python = ">=3.9"
urls = { homepage = "https://apexive.com/" }

dependencies = [
    "django",
    "django-rest-framework",
    "pilkit>=2.0",
    "django-imagekit>=4.1.0",
    # Pillow 10.0.0+ is incompatible with django-imagekit
    # https://github.com/matthewwithanm/django-imagekit/issues/566
    "Pillow<10",
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-django",
    "django-stubs",
    "mypy==1.5.0",
    "black==23.7.0",
    "ruff==0.0.284"
]

[build-system]
requires = ['setuptools>=68.1.0', 'wheel']
build-backend = 'setuptools.build_meta'

[tool.setuptools]
include-package-data = true
packages = ["df_cards"]


[tools.black]
max-line-length = 88


[tool.ruff]
line-length = 79
fix = true


# Enable Flake's "E" and "F" codes by default.
select = ["E", "F", "I", "W", "A", "B", "Q", "C", "S"]

ignore = ["B008", "E501", "S101", "A003"]

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    'static',
    'migrations',
    '__pycache__',
    '.pytest_cache',
    '__init__.py',
]
per-file-ignores = {}

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Assume Python 3.9.x
target-version = "py39"

[tool.mypy]

ignore_missing_imports = true
disable_error_code = "attr-defined, valid-type"
disallow_untyped_defs = true
mypy_path = "df_cards"
exclude = "venv|migrations|build|dist|docs"


[tool.pytest.ini_options]

python_files = "tests.py test_*.py"
DJANGO_SETTINGS_MODULE = "tests.settings"
